+++ .gitignore
@@ -11,6 +11,7 @@ node_modules
 types
 deployments
 .idea
+contracts-exposed
 
 # files
 *.env
