+++ test/token/ERC20/ERC20.test.ts
@@ -1,13 +1,13 @@
 const { BN, constants, expectEvent, expectRevert } = require('@openzeppelin/test-helpers');
 const { expect } = require('chai');
-const { ZERO_ADDRESS } = constants;
-
-const {
-  shouldBehaveLikeERC20,
-  shouldBehaveLikeERC20Transfer,
-  shouldBehaveLikeERC20Approve,
-} = require('./ERC20.behavior');
-const { expectRevertCustomError } = require('../../helpers/customError');
+// const { ZERO_ADDRESS } = constants;
+//
+// const {
+//   shouldBehaveLikeERC20,
+//   shouldBehaveLikeERC20Transfer,
+//   shouldBehaveLikeERC20Approve,
+// } = require('./ERC20.behavior');
+// const { expectRevertCustomError } = require('../../helpers/customError');
 
 const TOKENS = [
   { Token: artifacts.require('$ERC20') },
